review: standards compliance + tag the k6 setup health check
- Move the batched-enrichment test, mock method, and call recorder into
  entity_relation_service_test.go; extract seedVisitEnrichmentFixture so
  the test function is 60 lines (was 103, limit 100). Net change to the
  oversized entity_manager_test.go drops to +1 line (mock struct field).
- Wrap the buildAttrValuesAnchor error with attr/schema context instead
  of a bare return, and rephrase the new read-path errors to name the
  logical value type, the offending attribute/column, and the expected
  state per docs/error-handling.md.
- Tag the k6 setup() health check (qtype=setup_health_check) so every
  request in the script is attributable, matching the README claim.
